feat: rename hook for clarity
